pyproject.toml
[tool.poetry]
name = "webp2png converter"
version = "0.1.0"
description = ""
authors = ["ThanatosDi <yykkold55tw@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
